requirements.txt
tensorflow>=2.13.0
streamlit>=1.28.0
numpy>=1.23.0
pillow>=9.0.0
